Make travis build w/ and w/o requirements.txt
The reasoning being that when any other app installs lispify as a
dependency, pip doesn't look at the requirements.txt file at all.

See https://caremad.io/2013/07/setup-vs-requirement/ for an excellent
description of the different usage of a requirements.txt file and a
setup.py file. In short, setup.py lists abstract dependencies which pip
will install from PyPi or other sources listed in dependency_links and
is used during normal installation. requirements.txt on the other hand
is used during development.

It's important to test installing using one and the other separately
because it can help isolate bugs. requirements.txt is a list of
dependencies that are frozen during development when known to work, and
setup.py is a list of dependencies with perhaps some specification of
version for incompatibilities, but otherwise completely abstract and
versionless. The two lists can also diverge accidentally during
development, and if problems arise one can compare the results of
installing one way and the other to see what went wrong.